US#111Task#281 & 282 refactored so that all getCommits methods are present in one file, added new commit method to get total number of commits, refactored to get rid of cyclomatic complexity and refactored codeAnalysis to get rid of codacy token
